% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Function to add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{An integer/float.}

\item{y}{An integer/float.}
}
\value{
z Addition of the two numbers
}
\description{
Function to add two numbers
}
\examples{
add(5, 4)
}
