% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\name{filename}
\alias{filename}
\title{Function to modify a file name}
\usage{
filename(path)
}
\arguments{
\item{path}{A character string}
}
\value{
A character string
}
\description{
Function to modify a file name
}
